refactor: removed directly imports from plasmapy.utils
