+++ CITATION.cff
@@ -588,3 +588,6 @@ authors:
   family-names: Ahamed
   affiliation: OpenRefactory Inc.
   alias: fazledyn-or
+
+- given-names: Tomasz Adam
+  family-names: Skrzypczak
