+++ plasmapy/utils/__init__.py
@@ -2,36 +2,3 @@
 Package of functions and classes used to develop clean, readable, and informative
 code.
 """
-__all__ = [
-    "CouplingWarning",
-    "PhysicsError",
-    "PhysicsWarning",
-    "PlasmaPyError",
-    "PlasmaPyWarning",
-    "RelativityError",
-    "RelativityWarning",
-    "decorators",
-    "exceptions",
-    "roman",
-]
-
-import contextlib
-
-from plasmapy.utils import (
-    code_repr,
-    datatype_factory_base,
-    decorators,
-    exceptions,
-    roman,
-)
-from plasmapy.utils.exceptions import (
-    CouplingWarning,
-    PhysicsError,
-    PhysicsWarning,
-    PlasmaPyDeprecationWarning,
-    PlasmaPyError,
-    PlasmaPyFutureWarning,
-    PlasmaPyWarning,
-    RelativityError,
-    RelativityWarning,
-)
